langchain/requirement.txt
langchain
ipykernel
python-dotenv

langchain_community
pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
langchain-community
python-dotenv
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain>=0.2.0  # Or the latest version; avoid old ones like 0.1.5 if possible
langchain-community  # Required for community chains in 0.2+
langchain-core  # Core dependencies
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse_starlette